requirements.txt
fastapi
uvicorn
pydantic>=2.0
python-dotenv
sqlalchemy>=2.0
pyodbc
azure-storage-blob
azure-cosmos
openai>=1.0.0
pytest
httpx
